format moved to politikon templatetags
